chore: 방명록 수정 api 엔드포인트 수정
