+++ backend/src/main/kotlin/com/andlife/InvitationServer/controller/guestbook/GuestBookController.kt
@@ -1,0 +1,26 @@
+package com.andlife.InvitationServer.controller.guestbook
+
+import com.andlife.InvitationServer.request.invitation.guestbook.UpdateGuestBookRequest
+import com.andlife.InvitationServer.response.BaseResponse
+import com.andlife.InvitationServer.response.invitation.guestbook.GuestBookResponse
+import com.andlife.InvitationServer.service.invitation.guestbook.GuestBookService
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PutMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+@RequestMapping("/api/guestbooks")
+class GuestBookController(
+    private val guestBookService: GuestBookService,
+) {
+    @PutMapping("/{guestBookId}")
+    fun updateGuestBook(
+        @PathVariable guestBookId: Long,
+        @RequestBody request: UpdateGuestBookRequest
+    ): BaseResponse<GuestBookResponse> {
+        val result = guestBookService.updateGuestBook(guestBookId, request)
+        return BaseResponse.success(result)
+    }
+}
