+++ backend/src/main/kotlin/com/andlife/InvitationServer/request/invitation/guestbook/UpdateGuestBookRequest.kt
@@ -1,6 +1,6 @@
 package com.andlife.InvitationServer.request.invitation.guestbook
 
-data class GuestBookUpdateRequest(
+data class UpdateGuestBookRequest(
     val textContent: String,
     val existingImageIds: List<Long> = emptyList(),
     val existingVideoIds: List<Long> = emptyList(),
